### 2020-07-18 #2490 @nwu63
Hi @labarba, sorry about that I didn't realize it had to be a separate section. I've updated the paper and pushed it to the repo, please take a look if you have time. I would also be happy to suggest reviewers if needed.
