### 2020-07-20 2b53ae2
added more checks for linear constraints, and made sure they are not stored in conInfo
