### 2020-07-21 4f60b54
cast 1D array to 2D when single value is requested in getValues
