+++ pyoptsparse/pyOpt_history.py
@@ -626,6 +626,9 @@ class History(object):
         for name in names:
             # we just stack along axis 0
             data[name] = np.stack(data[name], axis=0)
+            # we cast 1D arrays to 2D, for scalar values
+            if data[name].ndim == 1:
+                data[name] = np.expand_dims(data[name], 1)
 
         return data
 
