### 2020-07-21 74aba6b
Update test_hs015.py
