### 2020-07-22 466aea3
Add entry point (#134)
* moved postprocessing into the repo

* added entry_point

* fixed relative import

* updated docs

* updated flake8 paths

* changed default directory to CWD
