### 2020-07-23 2756a61
added more checks for linear constraints, and made sure they are not stored in conInfo (#140)
