### 2020-07-24 12b7b05
skip scaling of linear constraints
