### 2020-07-24 c43089d
better way to remove linear constraints
