+++ pyoptsparse/pyOpt_history.py
@@ -257,10 +257,8 @@ class History(object):
         # only do this if we open the file with 'r' flag
         if self.flag != "r":
             return
-        conNames = list(self.conInfo.keys())
-        for con in conNames:
-            if self.optProb.constraints[con].linear:
-                conNames.remove(con)
+        # we remove linear constraints
+        conNames = [con for con in self.conInfo.keys() if not self.optProb.constraints[con].linear]
         return copy.deepcopy(conNames)
 
     def getObjNames(self):
@@ -433,7 +431,7 @@ class History(object):
         These are all "flat" dictionaries, with simple key:value pairs.
         """
         conDict = {}
-        for con in self.optProb.constraints.keys():
+        for con in list(self.optProb.constraints.keys()):
             # linear constraints are not stored in funcs
             if not self.optProb.constraints[con].linear:
                 conDict[con] = d["funcs"][con]
