### 2020-07-26 mdolab/pyoptsparse#145 @nwu63
Update testing framework
## Type of issue
What types of issue is it?
_Select the appropriate type(s) that describe this issue_

- Testing

## Description
We should update the overall testing framework:
- evaluate if we want to add other optimization problems
- share code among tests -- most of them are doing similar things and we should reduce code duplication
- improve overall code coverage
