### 2020-07-27 fe6e515
Additional linear constraint fixes (#144)
* skip scaling of linear constraints

* added test

* better way to remove linear constraints

* added comment
