### 2020-08-08 fc38786
added PSQP negative informs (#148)
