+++ pyoptsparse/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "2.1.4"
+__version__ = "2.1.5"
 
 from .pyOpt_history import History
 from .pyOpt_variable import Variable
