### 2020-08-11 #2490 @poulson
> Hi @poulson, just wondering if the paper can be moved to the review phase, or if it's waiting for some action from me.

This is on me. Thank you for your patience on this!
