### 2020-08-11 #2490 @poulson
Hi Severin (@sevberg) :wave: -- Given your expertise in Python and optimization, you would seem to be a great candidate for reviewing this submission on large-scale nonlinear optimization (in Python).

[FWIW: It seems that your username is incorrectly listed in the JOSS reviewer spreadsheet as @dsryberg]
