### 2020-08-11 #2490 @poulson
Hi David (@torressa) :wave: -- Given your expertise in Linear Programming and python, you seem like a great candidate for reviewing this submission on large-scale nonlinear optimization in Python.

Please let us know if you have the capacity and interest to review this software and paper.
