### 2020-08-11 #2490 @poulson
> hi, I would be available to review the paper and project but likely not before september

Thank you for the fast response, Mathieu. Both you and Vissarion (@vissarion) have the same constraint and would make a complete set of reviewers if added. I lean towards adding you both as reviewers and kicking off the review under the knowledge that the review won't start until September.
