### 2020-08-11 #2564 @whedon
[REVIEW]: pyOptSparse: A Python framework for large-scale constrained nonlinear optimization of sparse systems
**Submitting author:** @nwu63 (<a href="http://orcid.org/0000-0001-8856-9661">Neil Wu</a>)
**Repository:** <a href="https://github.com/mdolab/pyoptsparse/" target ="_blank">https://github.com/mdolab/pyoptsparse/</a>
**Version:** v2.2.1
