### 2020-08-14 39a49cb
fixed the IPOPT installation bug (#153)
