### 2020-08-17 #2564 @nwu63
Hi Mathieu, thanks for taking your time to do this review.

1. pyOptSparse is a wrapper for several popular optimizers, with a particular focus on large-scale nonlinear programming of continuous variables. As such, it is closer to frameworks such as pyomo and scipy, where the optimization problem is defined in a particular syntax and pyOptSparse performs the conversion to the specific form required for each optimizer. The main distinguishing feature for pyOptSparse is the support for sparse constraint Jacobians, both for linear and nonlinear constraints, and particularly for supporting sparse sub-blocks of the Jacobian. Through the unified framework, it's straightforward to switch between say SNOPT and IPOPT by changing one line of code, and all conversion will be done automatically including converting the sparse Jacobian into a different format. There are also other features such as the ability to perform finite-difference or complex-step to compute derivatives automatically, and MPI support for parallel function evaluations.

   The list of solvers is shown in the documentation (for example the side bar [here](https://mdolab-pyoptsparse.readthedocs-hosted.com/)). To switch between optimizers, the user would need to 1) change the options dictionary, and 2) change the name of the optimizer in the optimize call, for example replacing `SLSQP` with `IPOPT` in the line `opt = OPT("SLSQP", options=optOptions)`. I hesitate to provide a list and description of the optimizers in the paper because the list may change over time, with older optimizers eventually become deprecated and new ones added.

2. Yes, arbitrary Python functions are supported, and the data is stored in a dictionary. A nested dictionary is used for gradients if they are supplied. It's a little bit unclear for me what type of information should be in the paper, and what should be in the documentation site of the repository. I would prefer, if possible, to keep most of this information in the documentation since API may change and users typically will refer to that instead of the journal paper for help on using the package. But please let me know if JOSS takes a different view and I'd be happy to add more information to the paper. For reference, please see the [Quickstart](https://mdolab-pyoptsparse.readthedocs-hosted.com/en/latest/quickstart.html) for a basic example, and the [Guide](https://mdolab-pyoptsparse.readthedocs-hosted.com/en/latest/guide.html) for a complete example with sparse linear and nonlinear constraints.
