### 2020-08-27 #2564 @nwu63
Hi Mathieu, I agree with your points and I will push an update to the paper soon.
