### 2020-09-01 4f2aeb0
added some descriptions of features
