+++ paper/paper.md
@@ -9,6 +9,8 @@ authors:
     affiliation: 1
   - name: Gaetan Kenway
     affiliation: 1
+  - name: Charles A. Mader
+    affiliation: 1
   - name: John Jasa
     affiliation: 1
   - name: Joaquim R. R. A. Martins
@@ -43,6 +45,59 @@ $A$ is the linear constraint Jacobian, and $g(x)$ is the set of nonlinear constr
 pyOptSparse makes a distinction between linear and nonlinear constraints, since some optimizers have special treatments for linear constraints.
 Sparse linear constraints can be directly supplied in a number of different formats, and pyOptSparse will automatically handle the assembly and conversion for each optimizer.
 For nonlinear constraints, the sparsity pattern of the constraint Jacobian $\nabla g(x)$ can be specified as well.
+
+
+# Features
+## Support for multiple optimizers
+pyOptSparse provides built-in support for a number of popular proprietary and open-source optimizers.
+By isolating the optimization problem definition from the optimizer, the user is able to easily switch between different optimizers applied to the same optimization problem.
+In fact, a single line of code change is required, with minimal effort from the user.
+
+Although pyOptSparse is primarily focused on large-scale gradient-based optimization, it provides support for gradient-free optimizers as well.
+In addition, discrete variables, multi-objective and population-based optimizers are all supported.
+Due to the object-oriented programming approach, it is also straightforward to extend pyOptSparse to support any additional optimizers not currently available.
+That way, new optimizers would automatically inherit all the capabilities within pyOptSparse.
+
+## Assembly of sparse Jacobians
+
+## String-based indexing
+
+## Support for linear constraints
+
+## Automatic computation of derivatives
+If analytic derivatives for the objective and constraint functions are not available, pyOptSparse can automatically compute them internally using finite differences or complex step.
+For finite differences, the user has the option of either using forward or central differences, and with either an absolute or relative step size.
+Computing derivatives using finite differences can be expensive, requiring $n+1$ evaluations for forward differences and $2n+1$ for centered differences.
+They are also inaccurate due to subtractive cancellation error under finite precision arithmetic.
+Complex step [@Martins2003] on the other hand, still requires the same number of function evaluations.
+However, by perturbing each input along the imaginary axis, subtractive cancellation errors are avoided.
+As a result, by taking very small steps such as $10^{-40}i$, derivatives can be accurate down to machine precision.
+Of course, attention must be paid to ensure that the objective and constraint functions can be evaluated correctly with complex design variables.
+
+## Optimizer-independent problem scaling
+pyOptSparse offers optimizer-independent scaling for individual design variables, objective and constraints.
+By separating the optimization problem definition from the optimizer used, pyOptSparse is able to apply the scaling automatically and consistently to any chosen optimizer.
+Since the optimization problem is always defined in the physical, user-defined space, bounds on the design variables and constraints do not need to be modified when applying a different scaling.
+Furthermore, for gradient-based optimizers, all the derivatives are scaled automatically and consistently without any effort from the user.
+The only thing to specify is a `scale=` option when defining design variables, objective and constraints, and to change the scaling amounts to changing a single number.
+
+## Parallel execution
+pyOptSparse offers parallel execution in three ways.
+Firstly and most commonly, it can perform parallel function evaluations when the functions themselves require multiple processors.
+This is commonly the case when performing large-scale optimizations, where the objective and constraint functions are the result of a complex analysis such as computational fluid simulations.
+In this scenario, pyOptSparse can be executed with multiple processors, and all processors will be used to perform function evaluation, but only the root processor will run the optimizer itself.
+That way, we avoid the scenario where each processor is running an independent copy of the optimizer, and potentially causing inconsistencies or processor locking.
+
[…]
