+++ paper/paper.bib
@@ -2,7 +2,6 @@
     author = {Ruben E. Perez and Peter W. Jansen and Joaquim R. R. A. Martins},
     doi = {10.1007/s00158-011-0666-3},
     journal = {Structural and Multidisciplinary Optimization},
-    keywords = {optimization, Python},
     month = {January},
     number = {1},
     pages = {101--118},
@@ -41,7 +40,6 @@
 @article{Curtis2010,
     author = {Frank E. Curtis and Olaf Schenk and Adreas W\"achter},
     journal = {SIAM Journal on Scientific Computing},
-    keywords = {IPOPT, optimization},
     number = {6},
     pages = {3447--3475},
     title = {An Interior Point Algorithm for Large-Scale Nonlinear Optimization with Inexact Step Computations},
@@ -108,7 +106,6 @@
     author = {Ney R. Secco and Joaquim R. R. A. Martins},
     doi = {10.2514/1.C034934},
     journal = {Journal of Aircraft},
-    keywords = {xsede},
     month = {January},
     number = {1},
     pages = {217--227},
@@ -133,7 +130,6 @@
     author = {Justin S. Gray and John T. Hwang and Joaquim R. R. A. Martins and Kenneth T. Moore and Bret A. Naylor},
     doi = {10.1007/s00158-019-02211-z},
     journal = {Structural and Multidisciplinary Optimization},
-    keywords = {OpenMDAO},
     month = {April},
     number = {4},
     pages = {1075--1104},
@@ -195,4 +191,16 @@
     title = {The {NLopt} nonlinear-optimization package},
     url = {http://github.com/stevengj/nlopt},
     year = {2020},
+}
+
+@article{Martins2003,
+    author = {Joaquim R. R. A. Martins and Peter Sturdza and Juan J. Alonso},
+    doi = {10.1145/838250.838251},
+    journal = {ACM Transactions on Mathematical Software},
+    month = {September},
+    number = {3},
+    pages = {245--262},
+    title = {The Complex-Step Derivative Approximation},
+    volume = {29},
+    year = {2003}
 }
