+++ paper/paper.md
@@ -26,8 +26,8 @@ bibliography: paper.bib
 pyOptSparse is an optimization framework designed for constrained nonlinear optimization of large sparse problems, providing a unified interface for a variety of gradient-free and gradient-based optimizers.
 By using an object-oriented approach, it maintains independence between the optimization problem formulation and the implementation of the specific optimizers.
 The code is MPI-wrapped to enable execution of expensive parallel analyses and gradient evaluations, such as when using computational fluid dynamics (CFD) simulations which can require hundreds of processors.
-During the optimization, a history file can be stored to record the optimization history, which can be used both for post-processing and for performing an optimization hot-start.
-A graphical user interface application is also provided to interactively plot various quantities over an optimization.
+During the optimization, a history file can be stored to record the optimization history, which can be used both for post-processing and for restarting an optimization.
+A graphical user interface application is also provided to interactively visualize the optimization history.
 
 pyOptSparse considers optimization problems of the form
 \begin{align*}
@@ -42,15 +42,14 @@ g(x)\\
 \end{align*}
 where $x$ is the vector of design variables and $f(x)$ is a nonlinear objective function.
 $A$ is the linear constraint Jacobian, and $g(x)$ is the set of nonlinear constraint functions.
-pyOptSparse makes a distinction between linear and nonlinear constraints, since some optimizers have special treatments for linear constraints.
-Sparse linear constraints can be directly supplied in a number of different formats, and pyOptSparse will automatically handle the assembly and conversion for each optimizer.
-For nonlinear constraints, the sparsity pattern of the constraint Jacobian $\nabla g(x)$ can be specified as well.
 
 # Features
 ## Support for multiple optimizers
 pyOptSparse provides built-in support for a number of popular proprietary and open-source optimizers.
+Different optimizers typically have different requirements for formulating the optimization problem, such as constraint ordering, specification of equality constraints, and the sparse matrix format used to specify the constraint Jacobian.
+However, this is all handled automatically by the various optimizer wrappers.
 By isolating the optimization problem definition from the optimizer, the user is able to easily switch between different optimizers applied to the same optimization problem.
-In fact, a single line of code change is required, with minimal effort from the user.
+In fact, the user has to change only a single line of code to switch to another optimizer.
 
 Although pyOptSparse is primarily focused on large-scale gradient-based optimization, it provides support for gradient-free optimizers as well.
 In addition, discrete variables, multi-objective and population-based optimizers are all supported.
@@ -59,14 +58,15 @@ That way, new optimizers would automatically inherit all the capabilities within
 
 ## String-based indexing
 Unlike many of the other optimization frameworks out there, pyOptSparse is designed for large-scape optimizations, particularly in engineering applications.
-With up to $\mathcal{O}(1000)$ design variables and constraints, it is important to keep track of all values during computation.
+With $\mathcal{O}(1000)$ design variables and constraints, it is important to correctly keep track of all values during optimization.
 pyOptSparse employs string-based indexing to accomplish this.
-Instead of using a single flattened array, related design variables and constraints can be grouped together into arrays.
-These are combined into a Python dictionary, with each group identified by a unique key.
+Instead of using a single flattened array, related design variables and constraints can be grouped together into separate arrays.
+These are combined using an ordered dictionary, with each group identified by a unique key.
 Similarly, the constraint Jacobian is represented using a nested dictionary approach.
 This has several advantages:
+
 - Accessing design variable and constraint values without needing to consider their global indices, which is clumsy and prone to error.
[…]
