+++ pyoptsparse/postprocessing/OptViewDash/OptView_dash2.py
@@ -1,0 +1,1292 @@
+# This dash version of OptView makes use of the new API
+# to read in the history file, rather than using the
+# OptView baseclass. This should be more maintainable
+# for adding new features or displaying new information with OptView.
+
+# !/usr/bin/python
+import dash
+import dash_core_components as dcc
+import dash_html_components as html
+import plotly.graph_objs as go
+from plotly import subplots
+import numpy as np
+import argparse
+import sys
+from pyoptsparse import History
+import json
+
+# Read in the history files given by user
+major_python_version = sys.version_info[0]
+parser = argparse.ArgumentParser()
+parser.add_argument(
+    "histFile", nargs="*", type=str, default="opt_hist.hst", help="Specify the history file to be plotted"
+)
+args = parser.parse_args()
+
+# List of strings (the history file names)
+histListArgs = args.histFile
+
+# Create list of associated labels (A, B, C, etc) for each history file if there are more than one
+index = ord("A")
+fileLabels = ["" for file in histListArgs]
+if len(fileLabels) > 1:
+    for i in range(len(fileLabels)):
+        fileLabels[i] = chr(index)
+        index = (index + 1 - 65) % 26 + 65
+
+# Color scheme for graphing traces
+colors = ["#636EFA", "#EF553B", "#00CC96", "#AB63FA", "#FFA15A", "#19D3F3", "#FF6692", "#B6E880", "#FF97FF", "#FECB52"]
+
+
+# ====================================================================================
+#  Defining dash app & layout
+# ====================================================================================
+
+app = dash.Dash(__name__)
+# Override expceptions for when elements are defined without initial input
+app.config.suppress_callback_exceptions = True
+
+app.layout = html.Div(
+    children=[
+        html.Nav(
+            children=[
+                html.Img(
+                    src=app.get_asset_url("/logo.png"), style={"height": "4rem", "width": "13rem", "margin": "0.2rem"}
+                ),
+                html.Div(["OptView"], style={"fontWeight": "550", "fontSize": "3.0rem", "color": "#193D72"}),
+            ],
+            style={"display": "flex", "backgroundColor": "#EEEEEE"},
+        ),
+        html.Div(
+            [
+                html.Div(
+                    [
+                        html.Div(
+                            [
+                                html.Div(
+                                    [
+                                        html.Div(
+                                            ["History Files"],
+                                            style={"color": "#193D72", "fontWeight": "normal", "fontSize": "15px"},
+                                        ),
+                                        html.Ul(
+                                            [
+                                                html.Li(
+                                                    [fileLabels[i] + " " + histListArgs[i]], style={"fontSize": "12px"}
+                                                )
+                                                for i in range(len(histListArgs))
+                                            ]
+                                        ),
+                                    ],
+                                ),
+                                html.Div(
+                                    [
+                                        html.Div(
+                                            ["Design Groups"],
+                                            className="dvarGroup",
+                                            style={"color": "#193D72", "fontWeight": "normal", "fontSize": "1.5rem"},
+                                        ),
+                                        dcc.Dropdown(
+                                            id="dvarGroup",
+                                            placeholder="Select design group(s)...",
+                                            multi=True,
+                                            style={"color": "#676464", "fontSize": "1.2rem"},
+                                        ),
+                                    ],
+                                ),
+                                html.Div(
+                                    [
+                                        html.Div(
+                                            ["Design Variables"],
+                                            style={"color": "#193D72", "fontWeight": "normal", "fontSize": "1.5rem"},
+                                        ),
+                                        dcc.Dropdown(
+                                            id="dvarChild",
+                                            placeholder="Select design var(s)...",
+                                            multi=True,
+                                            style={"color": "#676464", "fontSize": "1.2rem"},
+                                        ),
+                                    ],
+                                ),
+                            ],
[…]
