+++ pyoptsparse/postprocessing/OptViewDash/requirements.txt
@@ -1,0 +1,2 @@
+dash
+plotly
