### 2020-09-08 415de7e
Documented  addMaxMinTraces
