### 2020-09-09 d9d6eec
Edited whole paper. Please check diffs.
