+++ paper/paper.md
@@ -23,11 +23,11 @@ bibliography: paper.bib
 ---
 
 # Summary
-pyOptSparse is an optimization framework designed for constrained nonlinear optimization of large sparse problems, providing a unified interface for a variety of gradient-free and gradient-based optimizers.
-By using an object-oriented approach, it maintains independence between the optimization problem formulation and the implementation of the specific optimizers.
-The code is MPI-wrapped to enable execution of expensive parallel analyses and gradient evaluations, such as when using computational fluid dynamics (CFD) simulations which can require hundreds of processors.
-During the optimization, the user can store the optimization history in a database file, which can then be used both for post-processing and for restarting an optimization.
-A graphical user interface application is also provided to interactively visualize the optimization history.
+pyOptSparse is an optimization framework designed for constrained nonlinear optimization of large sparse problems, providing a unified interface for various gradient-free and gradient-based optimizers.
+Using an object-oriented approach maintains independence between the optimization problem formulation and the implementation of the specific optimizers.
+The code is MPI-wrapped to enable execution of expensive parallel analyses and gradient evaluations, such as when using computational fluid dynamics (CFD) simulations, which can require hundreds of processors.
+The optimization history can be stored in a database file, which can then be used both for post-processing and restarting another optimization.
+A graphical user interface application is provided to visualize the optimization history interactively.
 
 pyOptSparse considers optimization problems of the form
 \begin{align*}
@@ -45,94 +45,96 @@ $A$ is the linear constraint Jacobian, and $g(x)$ is the set of nonlinear constr
 
 # Features
 ## Support for multiple optimizers
-pyOptSparse provides built-in support for a number of popular proprietary and open-source optimizers.
-Different optimizers typically have different requirements for formulating the optimization problem, such as constraint ordering, specification of equality constraints, and the sparse matrix format used to specify the constraint Jacobian.
-However, thankfully this is all handled automatically by the various optimizer wrappers within pyOptSparse.
-By isolating the optimization problem definition from the optimizer, the user is able to easily switch between different optimizers applied to the same optimization problem.
-In fact, the user has to change only a single line of code to switch to another optimizer.
+pyOptSparse provides built-in support for several popular proprietary and open-source optimizers.
+Each optimizer usually has its own way to specify the problem:
+It might require different constraint ordering, have different ways of specifying equality constraints, or use a sparse matrix format to represent the constraint Jacobian.
+pyOptSparse provides a common Python interface for the various optimizer that hides these differences from the user. 
+By isolating the optimization problem definition from the optimizer, the user can easily switch between different optimizers applied to the same optimization problem.
+The optimizer can be switched by editing a single line of code.
 
-Although pyOptSparse is primarily focused on large-scale gradient-based optimization, it provides support for gradient-free optimizers as well.
-In addition, discrete variables, multi-objective and population-based optimizers are all supported.
-Due to the object-oriented programming approach, it is also straightforward to extend pyOptSparse to support any additional optimizers not currently available.
-That way, new optimizers would automatically inherit all the capabilities within pyOptSparse.
+Although pyOptSparse focuses primarily on large-scale gradient-based optimization, it provides support for gradient-free optimizers as well.
+Also, discrete variables, multi-objective, and population-based optimizers are all supported.
+Because of the object-oriented programming approach, it is also straightforward to extend pyOptSparse to support any additional optimizers that are not currently available.
+When adding a new optimizer, it automatically inherits all the capabilities within pyOptSparse.
+[//]: # (TODO JM-: What do you mean by this last sentence? The optimizer does not gain any new capability, just a new interface, right?)
 
 ## String-based indexing
-Unlike many of the other publicly available optimization frameworks, pyOptSparse is designed for large-scale optimizations, with a particular focus on engineering applications.
[…]
