+++ test/test_snopt_bugfix.py
@@ -53,8 +53,7 @@ def objfunc_2con(xdict):
 
 
 def sens(xdict, funcs):
-    """f(x,y) = (x-3)^2 + xy + (y+4)^2 - 3
-    """
+    """f(x,y) = (x-3)^2 + xy + (y+4)^2 - 3"""
     x = xdict["x"]
     y = xdict["y"]
     funcsSens = {}
