+++ pyoptsparse/pyIPOPT/pyIPOPT.py
@@ -413,7 +413,7 @@ class IPOPT(Optimizer):
         storeSens : bool
             Flag sepcifying if sensitivities are to be stored in hist.
             This is necessay for hot-starting only.
-            """
+        """
 
         self.callCounter = 0
         self.storeSens = storeSens
