+++ examples/snopt_bugfix.py
@@ -23,8 +23,7 @@ def objfunc(xdict):
 
 
 def sens(xdict, funcs):
-    """f(x,y) = (x-3)^2 + xy + (y+4)^2 - 3
-    """
+    """f(x,y) = (x-3)^2 + xy + (y+4)^2 - 3"""
     x = xdict["x"]
     y = xdict["y"]
     funcsSens = {}
