+++ pyoptsparse/pyParOpt/ParOpt.py
@@ -122,7 +122,7 @@ class ParOpt(Optimizer):
         storeSens : bool
             Flag sepcifying if sensitivities are to be stored in hist.
             This is necessay for hot-starting only.
-            """
+        """
 
         self.callCounter = 0
         self.storeSens = storeSens
