+++ pyoptsparse/pyCONMIN/pyCONMIN.py
@@ -110,7 +110,7 @@ class CONMIN(Optimizer):
         storeSens : bool
             Flag sepcifying if sensitivities are to be stored in hist.
             This is necessay for hot-starting only.
-            """
+        """
 
         self.callCounter = 0
         self.storeSens = storeSens
