+++ pyoptsparse/pyPSQP/pyPSQP.py
@@ -126,7 +126,7 @@ class PSQP(Optimizer):
         storeSens : bool
             Flag sepcifying if sensitivities are to be stored in hist.
             This is necessay for hot-starting only.
-            """
+        """
 
         self.callCounter = 0
         self.storeSens = storeSens
