### 2020-09-10 231d984
Some updates, mostly grammar (#158)
