+++ paper/paper.md
@@ -23,7 +23,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-pyOptSparse is an optimization framework designed for constrained nonlinear optimization of large sparse problems, providing a unified interface for various gradient-free and gradient-based optimizers.
+pyOptSparse is an optimization framework designed for constrained nonlinear optimization of large sparse problems and provides a unified interface for various gradient-free and gradient-based optimizers.
 Using an object-oriented approach maintains independence between the optimization problem formulation and the implementation of the specific optimizers.
 The code is MPI-wrapped to enable execution of expensive parallel analyses and gradient evaluations, such as when using computational fluid dynamics (CFD) simulations, which can require hundreds of processors.
 The optimization history can be stored in a database file, which can then be used both for post-processing and restarting another optimization.
@@ -48,15 +48,16 @@ $A$ is the linear constraint Jacobian, and $g(x)$ is the set of nonlinear constr
 pyOptSparse provides built-in support for several popular proprietary and open-source optimizers.
 Each optimizer usually has its own way to specify the problem:
 It might require different constraint ordering, have different ways of specifying equality constraints, or use a sparse matrix format to represent the constraint Jacobian.
-pyOptSparse provides a common Python interface for the various optimizer that hides these differences from the user. 
+pyOptSparse provides a common Python interface for the various optimizers that hides these differences from the user. 
 By isolating the optimization problem definition from the optimizer, the user can easily switch between different optimizers applied to the same optimization problem.
 The optimizer can be switched by editing a single line of code.
 
 Although pyOptSparse focuses primarily on large-scale gradient-based optimization, it provides support for gradient-free optimizers as well.
 Also, discrete variables, multi-objective, and population-based optimizers are all supported.
 Because of the object-oriented programming approach, it is also straightforward to extend pyOptSparse to support any additional optimizers that are not currently available.
-When adding a new optimizer, it automatically inherits all the capabilities within pyOptSparse.
+All of the capabilities within pyOptSparse, including history recording and postprocessing, are automatically inherited when new optimizers are added.
 <!-- [//]: (TODO JM-: What do you mean by this last sentence? The optimizer does not gain any new capability, just a new interface, right?) -->
+<!-- JJ rephrased sentence, feel free to fine-tune -->
 
 ## String-based indexing
 Unlike many other publicly available optimization frameworks, pyOptSparse is designed to handle large-scale optimizations, focusing on engineering applications.
@@ -67,10 +68,10 @@ These arrays are combined using an ordered dictionary, where each group is ident
 Similarly, the constraint Jacobian is represented by a nested dictionary approach.
 This representation has several advantages:
 
-- The design variable and constraint values can be accessed without needing to consider their global indices, which is clumsy and prone to error.
-- The global indices are also often optimizer-dependent, and this extra level of abstraction allows pyOptSparse to handle everything behind the scenes.
+- The design variable and constraint values can be accessed without knowing their global indices, which reduces possible user error.
+- The global indices are also often optimizer-dependent and this extra level of wrapping abstracts away potentially confusing differences between optimizers.
 - The constraint Jacobian can be computed and provided at the sub-block level, leaving pyOptSparse to assemble the whole Jacobian.
-  This mimics the engineering workflow, where different tools often compute different sub-blocks of the Jacobian.
+  This mimics the engineering workflow where different tools often compute different sub-blocks of the Jacobian.
   The user only has to ensure that the indices within each sub-block are correct, and the rest is handled automatically.
 
 ## Support for sparse linear and nonlinear constraints
@@ -79,7 +80,7 @@ When defining constraints, it is possible to provide the sparsity pattern of the
[…]
