### 2020-09-10 #2564 @vissarion
I finished my review.

I agree with the points listed above. I would also like to see a high-level example probably with the comment that links to the documentation for more details.

One more comment, I think it would be useful to note in the paper the current release version of the package.
