+++ paper/paper.bib
@@ -204,3 +204,10 @@
     volume = {29},
     year = {2003}
 }
+
+@techreport{Kraft1988a,
+    author = {Kraft, D},
+    institution = {Tech. Rep. DFVLR-FB 88-28, DLR German Aerospace Center},
+    title = {{A software package for sequential quadratic programming}},
+    year = {1988}
+}
