### 2020-09-10 #2564 @nwu63
Hi Mathieu and Vissarion,

Thanks for your feedback and apologies for not replying sooner, but I have been working with co-authors to update the paper. There are some substantial changes:
- I have added a `Features` section to explain in more detail some salient features of the software which are not found elsewhere
- I have added a section containing an example optimization script. There is no guarantee that the public API would remain stable, but I do think it's useful in showing what an optimization script would look like.
- I have addressed other, more minor comments

The paper is located under the `paper` branch of the repo, please take a look and let me know if you have any feedback.

Cheers,
Neil
