### 2020-09-10 f9a8629
added extra install dependencies
