+++ doc/requirements.txt
@@ -1,1 +1,1 @@
-numpydoc
+-e .[docs]
