### 2020-09-10 8959fc4
Removed patch version from dependencies (#160)
