### 2020-09-10 mdolab/pyoptsparse#145 @matbesancon
Linking the review issue:
https://github.com/mdolab/pyoptsparse/issues/145
