### 2020-09-10 #2564 @matbesancon
Posted on https://github.com/mdolab/pyoptsparse/issues/145 for the testing part which seems to have issues
