### 2020-09-10 mdolab/pyoptsparse#145 @nwu63
Hi Mathieu, would you mind opening a new issue for this so that we may discuss there? This issue is more about the overall scope of testing, and refactoring some existing tests.
