### 2020-09-10 mdolab/pyoptsparse#145 @matbesancon
Deleted comment that was my mistake
