### 2020-09-10 #2564 @matbesancon
Ok nevermind the tests do work. On the installation instructions [here](https://mdolab-pyoptsparse.readthedocs-hosted.com/en/latest/install.html), I would recommend you state upfront that a fortran compiler is needed (maybe recommend gfortran directly)
