### 2020-09-10 #2564 @nwu63
I've updated the installation instructions [in this PR](https://github.com/mdolab/pyoptsparse/pull/161).
