### 2020-09-12 9e8446e
added test for unconstrained opt - Rosenbrock
