+++ test/test_rosenbrock.py
@@ -1,0 +1,249 @@
+"""Test solution of Rosenbrock problem"""
+
+import unittest
+import numpy as np
+from numpy.testing import assert_allclose
+from pyoptsparse import Optimization, OPT, History
+from pyoptsparse.pyOpt_error import Error
+
+
+class TestRosenbrock(unittest.TestCase):
+
+    ## Solve unconstrained Rosenbrock problem.
+    #  This problem is scalable w.r.t. design variables number.
+    #  We select a problem with 4 design variables, but the
+    #  location and value of the minimum do not change with DV
+    #  dimensionality
+    #
+    #
+    #  min   100 * (x[i + 1] - x[i] ** 2) ** 2 + (1 - x[i]) ** 2
+    #
+    #  The minimum is located at x=(1,....,1) where x
+    #  is an arbitrarily sized vector depending on the number N
+    #  of design variables.
+    #  At the optimum, the function is f(x) = 0.
+    #  We select a random initial point for our test.
+    ##
+
+    def objfunc(self, xdict):
+        self.nf += 1
+        x = xdict["xvars"]
+
+        funcs = {}
+        funcs["obj"] = 0
+
+        for i in range(len(x) - 1):
+            funcs["obj"] += 100 * (x[i + 1] - x[i] ** 2) ** 2 + (1 - x[i]) ** 2
+
+        fail = False
+        return funcs, fail
+
+    def sens(self, xdict, funcs):
+        self.ng += 1
+        x = xdict["xvars"]
+        funcsSens = {}
+        grads = np.zeros(len(x))
+
+        for i in range(len(x) - 1):
+            grads[i] += 2 * (200 * x[i] ** 3 - 200 * x[i] * x[i + 1] + x[i] - 1)
+            grads[i + 1] += 200 * (x[i + 1] - x[i] ** 2)
+
+        funcsSens["obj"] = {"xvars": grads}
+
+        fail = False
+        return funcsSens, fail
+
+    def optimize(self, optName, tol, optOptions={}, storeHistory=False, hotStart=None):
+        self.nf = 0  # number of function evaluations
+        self.ng = 0  # number of gradient evaluations
+        # Optimization Object
+
+        optProb = Optimization("Rosenbrock Problem", self.objfunc)
+
+        n = 4  # Number of design variables
+        np.random.seed(10)
+        value = np.random.normal(size=n)
+
+        lower = np.ones(n) * -50
+        upper = np.ones(n) * 50
+        optProb.addVarGroup("xvars", n, lower=lower, upper=upper, value=value)
+
+        # Objective
+        optProb.addObj("obj")
+
+        # Check optimization problem:
+        print(optProb)
+
+        # Optimizer
+        try:
+            opt = OPT(optName, options=optOptions)
+        except Error:
+            raise unittest.SkipTest("Optimizer not available:", optName)
+
+        # Solution
+        if storeHistory is not None:
+            if storeHistory is True:
+                self.histFileName = "%s_Rsbrk_Hist.hst" % (optName.lower())
+            elif isinstance(storeHistory, str):
+                self.histFileName = storeHistory
+        else:
+            self.histFileName = None
+
+        sol = opt(optProb, sens=self.sens, storeHistory=self.histFileName, hotStart=hotStart)
+
+        # Test printing solution to screen
+        print(sol)
+
+        # Check Solution
+        self.fStar1 = 0.0
+
+        self.xStar1 = np.ones(n)
+
+        dv = sol.getDVs()
+        sol_xvars = [sol.variables["xvars"][i].value for i in range(n)]
+
+        assert_allclose(sol_xvars, dv["xvars"], atol=tol, rtol=tol)
+        # we check either optimum via try/except
+        # try:
[…]
