### 2020-09-12 443051f
fixed NLPQLP for unconstrained problems
