+++ test/test_rosenbrock.py
@@ -103,12 +103,9 @@ class TestRosenbrock(unittest.TestCase):
         sol_xvars = [sol.variables["xvars"][i].value for i in range(n)]
 
         assert_allclose(sol_xvars, dv["xvars"], atol=tol, rtol=tol)
-        # we check either optimum via try/except
-        # try:
+
         assert_allclose(sol.objectives["obj"].value, self.fStar1, atol=tol, rtol=tol)
         assert_allclose(dv["xvars"], self.xStar1, atol=tol, rtol=tol)
-        # except AssertionError:
-        #     pass
 
     def check_hist_file(self, optimizer, tol):
         """
