### 2020-09-12 67146ff
removed unused checks
