### 2020-09-12 dffc4cb
Bugfix and testing for unconstrained optimization (#162)
