### 2020-09-22 8ac5537
removed dependency 'six'
