+++ test/test_snopt_user_termination.py
@@ -40,10 +40,10 @@ class TerminateComp(object):
         """f(x,y) = (x-3)^2 + xy + (y+4)^2 - 3"""
         x = xdict["x"]
         y = xdict["y"]
-        funcsSens = {}
+        funcsSens = {"obj": {}}
 
-        funcsSens["obj", "x"] = 2.0 * x - 6.0 + y
-        funcsSens["obj", "y"] = 2.0 * y + 8.0 + x
+        funcsSens["obj"]["x"] = 2.0 * x - 6.0 + y
+        funcsSens["obj"]["y"] = 2.0 * y + 8.0 + x
 
         if self.sens_count > self.max_sens:
             fail = 2
