### 2020-09-22 a24cc6c
removed other unused code
