### 2020-09-22 3bcbef2
added back bashrc_mdolab since it's an executable
