+++ .travis.yml
@@ -21,9 +21,8 @@ env:
     - DOCKER_TAG=c7-intel-impi-latest
     - DOCKER_TAG=u18-gcc-ompi-stable
     - DOCKER_TAG=u18-gcc-ompi-latest
-    - DOCKER_TAG=u18-gcc-ompi-stable
-    - DOCKER_TAG=u20-gcc-ompi-latest
     - DOCKER_TAG=u20-gcc-ompi-stable
+    - DOCKER_TAG=u20-gcc-ompi-latest
 
 before_install:
   - if [ "$TRAVIS_PULL_REQUEST" != "false" ]; then
