### 2020-09-23 d344c12
Remove python 2 support (#164)
* updated travis

* removed dependency 'six'

* updated docs

* updated github actions

* added back bashrc_mdolab since it's an executable

* Update .travis.yml
