### 2020-09-23 4eba2ec
updated travis and made nested dicts look better
