+++ .travis.yml
@@ -54,7 +54,7 @@ install:
       docker exec -it app /bin/bash -c "cp -r \$HOME/NLPQLP/* $NLPQLP_DIR/ && cp -r \$HOME/SNOPT/* $SNOPT_DIR/";
     fi
   # Build and install
-  - docker exec -it app /bin/bash -c ". \$HOME/.bashrc && cd $DOCKER_WORKING_DIR && pip install -e ."
+  - docker exec -it app /bin/bash -c ". \$HOME/.bashrc_mdolab && cd $DOCKER_WORKING_DIR && pip install -e ."
 script:
   # We need to source the mdolab bashrc before running anything
   - docker exec -it app /bin/bash -c ". \$HOME/.bashrc_mdolab && cd $DOCKER_WORKING_DIR && testflo --pre_announce -v . --coverage --coverpkg $REPO_NAME"
@@ -63,6 +63,6 @@ after_success:
   # Coveralls
   # These environment variables must be exposed to coveralls for it to work (since we are on Docker)
   - if [ "$TRAVIS_PULL_REQUEST" == "false" ]; then
-      docker exec -it app /bin/bash -c ". \$HOME/.bashrc && pip install coveralls";
+      docker exec -it app /bin/bash -c ". \$HOME/.bashrc_mdolab && pip install coveralls";
       docker exec -e COVERALLS_REPO_TOKEN="$COVERALLS_REPO_TOKEN" -e TRAVIS_JOB_ID="$TRAVIS_JOB_ID" -e TRAVIS_BRANCH="$TRAVIS_BRANCH" -it app /bin/bash -c ". \$HOME/.bashrc && cd $DOCKER_WORKING_DIR && coveralls --rcfile=.coveragerc";
     fi
