### 2020-09-23 87939d6
Remove flat sensitivity dict (#165)
* updated travis

* removed dependency 'six'

* removed flat sens dict

* removed other unused code

* Update .travis.yml

* updated travis and made nested dicts look better

* Update .travis.yml

* fixed travis
