+++ .travis.yml
@@ -3,6 +3,7 @@ import:
 
 env:
   global:
+    - DOCKER_WORKING_DIR=/home/mdolabuser/packages/$REPO_NAME # pyOptSparse is under packages, so this should overwrite what's set in travis-env.yml
     - SNOPT_DIR=$DOCKER_WORKING_DIR/$REPO_NAME/pySNOPT/source
     - NLPQLP_DIR=$DOCKER_WORKING_DIR/$REPO_NAME/pyNLPQLP/source
 
