### 2020-09-27 1acadf3
Update Travis (#167)
* Update .travis.yml

* Update .travis.yml
