+++ .travis.yml
@@ -1,29 +1,11 @@
-os:
-- linux
+import:
+  - mdolab/.github:travis-env.yml
 
-branches:
-  only:
-  - master
-
-language: generic
-
-services:
-  - docker
 env:
   global:
-    - REPO_NAME=$(basename $TRAVIS_REPO_SLUG)
-    - PR_FROM=$(dirname $TRAVIS_PULL_REQUEST_SLUG)
-    - DOCKER_WORKING_DIR=/home/mdolabuser/packages/$REPO_NAME
-    - DOCKER_MOUNT_DIR=/home/mdolabuser/travis/$REPO_NAME
+    - DOCKER_WORKING_DIR=/home/mdolabuser/packages/$REPO_NAME # pyOptSparse is under packages, so this should overwrite what's set in travis-env.yml
     - SNOPT_DIR=$DOCKER_WORKING_DIR/$REPO_NAME/pySNOPT/source
     - NLPQLP_DIR=$DOCKER_WORKING_DIR/$REPO_NAME/pyNLPQLP/source
-  jobs:
-    - DOCKER_TAG=c7-gcc-ompi-latest
-    - DOCKER_TAG=c7-intel-impi-latest
-    - DOCKER_TAG=u18-gcc-ompi-latest
-    - DOCKER_TAG=u18-gcc-ompi-stable
-    - DOCKER_TAG=u20-gcc-ompi-latest
-    - DOCKER_TAG=u20-gcc-ompi-stable
 
 before_install:
   - if [ $PR_FROM != "mdolab" ]; then
@@ -39,7 +21,6 @@ before_install:
         --mount "type=bind,src=$(pwd),target=$DOCKER_MOUNT_DIR"
         mdolab/$DOCKER_REPO:$DOCKER_TAG
         /bin/bash
-  # set envars for coveralls inside docker
 
 install:
   # We back up the proprietary source codes first
@@ -55,15 +36,15 @@ install:
       docker exec -it app /bin/bash -c "cp -r \$HOME/NLPQLP/* $NLPQLP_DIR/ && cp -r \$HOME/SNOPT/* $SNOPT_DIR/";
     fi
   # Build and install
-  - docker exec -it app /bin/bash -c ". \$HOME/.bashrc_mdolab && cd $DOCKER_WORKING_DIR && pip install -e ."
+  - docker exec -it app /bin/bash -c ". $BASHRC && cd $DOCKER_WORKING_DIR && pip install -e ."
 script:
   # We need to source the mdolab bashrc before running anything
-  - docker exec -it app /bin/bash -c ". \$HOME/.bashrc_mdolab && cd $DOCKER_WORKING_DIR && testflo --pre_announce -v . --coverage --coverpkg $REPO_NAME"
+  - docker exec -it app /bin/bash -c ". $BASHRC && cd $DOCKER_WORKING_DIR && testflo --pre_announce -v . --coverage --coverpkg $REPO_NAME"
 
 after_success:
   # Coveralls
   # These environment variables must be exposed to coveralls for it to work (since we are on Docker)
   - if [ "$TRAVIS_PULL_REQUEST" == "false" ]; then
-      docker exec -it app /bin/bash -c ". \$HOME/.bashrc_mdolab && pip install coveralls";
+      docker exec -it app /bin/bash -c ". $BASHRC && pip install coveralls";
       docker exec -e COVERALLS_REPO_TOKEN="$COVERALLS_REPO_TOKEN" -e TRAVIS_JOB_ID="$TRAVIS_JOB_ID" -e TRAVIS_BRANCH="$TRAVIS_BRANCH" -it app /bin/bash -c ". \$HOME/.bashrc_mdolab && cd $DOCKER_WORKING_DIR && coveralls --rcfile=.coveragerc";
     fi
