+++ .travis.yml
@@ -1,6 +1,7 @@
 import:
   - mdolab/.github:travis/jobs-env.yml
-  - mdolab/.github:travis/before-install.yml
+  - source: mdolab/.github:travis/before-install.yml
+    mode: deep_merge_prepend
   - mdolab/.github:travis/script.yml
 
 env:
