+++ .travis.yml
@@ -1,7 +1,7 @@
 import:
   - mdolab/.github:travis/jobs-env.yml
   - source: mdolab/.github:travis/before-install.yml
-    mode: deep_merge_append
+    mode: deep_merge_prepend
   - mdolab/.github:travis/script.yml
 
 env:
@@ -9,7 +9,6 @@ env:
     - DOCKER_WORKING_DIR=/home/mdolabuser/packages/$REPO_NAME # pyOptSparse is under packages, so this should overwrite what's set in travis-env.yml
     - SNOPT_DIR=$DOCKER_WORKING_DIR/$REPO_NAME/pySNOPT/source
     - NLPQLP_DIR=$DOCKER_WORKING_DIR/$REPO_NAME/pyNLPQLP/source
-    - DOCKER_REPO=public
 
 before_install:
   - if [ $PR_FROM != "mdolab" ]; then
