+++ .travis.yml
@@ -34,7 +34,9 @@ install:
   - docker exec -it app /bin/bash -c ". $BASHRC && cd $DOCKER_WORKING_DIR && pip install ."
 
 script:
-  - if [ $TEST_TYPE == "docker" ]; then
+  - |
+    if [ $TEST_TYPE == "docker" ]; then
+      # run tests
       docker exec -it app /bin/bash -c ". $BASHRC && cd $DOCKER_WORKING_DIR && testflo --pre_announce -v . --coverage --coverpkg $REPO_NAME";
     fi
 
