### 2020-10-05 mdolab/pyoptsparse#169 @matbesancon
Replace string option for solver with constant or type
In the solver definition
opt = OPT(opt_name, options), having a string with the name of the solver is error-prone (a typo in the solver name) and hard to discover. Furthermore, it makes the integration of other solvers hard since a list of string names has to be maintained in sync with the solvers.
Passing a solver object (for example from a solvers module) could make things easier on the user.

Issue written for the JOSS review https://github.com/openjournals/joss-reviews/issues/2564
