### 2020-10-05 #2564 @matbesancon
Regarding the paper, thanks to the authors for the modifications. It is much clearer as-is. Few additional comments and remarks:

- The key motivation for the software is to handle sparse structures. Could the authors provide an idea of the size of problems where the sparse structure pays off compared to representing the problem with dense data structures (vectors and matrices)? Have some comparisons been made with PyOpt?
- Leveraging the history file seems like an interesting feature. It is hard to understand from the paragraph how the information is represented and what this file consists of. Is it a list of the previously-solved problems and their solutions? Is it automatically storing all problems or does the user have to store a problem and solution themselves?

These are the last points I consider worth expanding before the paper is ready
