### 2020-10-05 #2564 @nwu63
Hi Mathieu, thanks for the detailed review, I'll address the code issues shortly. As for your questions:
- The motivation is not for data storage or computational efficiency in numerical linear algebra, since we only deal with problems with maybe a few thousand DVs. Obviously there would be benefits if you are using hundreds of thousands of DVs, but this is not typically done in our work. However, the benefits of using a sparse Jacobian is two-fold:
  - First, it tells the optimizer which entries of the Jacobian will always be zero. Some optimizers may leverage this fact---for all function evaluations these gradient terms are going to be zero---during optimization. Certain optimizers may also internally benefit from such sparse formats in computation/storage as you mentioned, but this would be optimizer-specific. If an optimizer is able to accept sparse Jacobians, we try to provide it.
  - Second, the block-sparse format in pyOptSparse is very useful in engineering settings. pyOptSparse allows for the Jacobian to be computed at a sub-block level, and those not computed are automatically assumed to be zero. This means different engineering tools can separately compute different parts of the Jacobian, and pyOptSparse will assemble everything together automatically, assuming the rest of the entries are all zero. This naturally motivates the use of a sparse format for the Jacobian.
- The history file is a database that contains the iteration history for a previous optimization. It stores the design points queried, the function outputs (objective and constraint values), along with some metadata about the optimization such as the optimizer options. This can be used for optimization hotstart or visualization. The file is stored using the `storeHistory=` keyword argument when starting an optimization, for example `opt(optProb, sens='FD', storeHistory='my_history.hst')`.

Please let me know if this was not clear from the paper, and how I can improve it further.

Thanks again,
Neil
