### 2020-10-05 #2564 @matbesancon
Thanks for your answer. For the motivation, I feel the first point (telling the optimizer which entries of the Jacobian will always be zero), this is a performance reason, hence my question of comparing it to a dense formulation or modelling library.
In any case, the motivation could be expanded a bit in the statement of need section.

The use case for the history (and the way it works) could also be more detailed in the paper, otherwise the reader might not understand what this implies for them as users
