### 2020-10-05 #2564 @nwu63
- Right, I agree. However I just want to point out that we do not implement anything in the optimizer -- all the optimizers here are developed by others. We merely provide sparse Jacobians for optimizers that support them, so the discussion on performance would be specific to each optimizer, and IMO outside the scope of this work. The main contribution here is related to the second point, where the construction of the sparse Jacobian at a sub-block level is tailored for engineering applications.
- Yes I agree.

I'll update the paper, as well as address the opened issues. Thanks again!
