### 2020-10-05 mdolab/pyoptsparse#169 @nwu63
It's currently possible to instantiate a solver object without using strings, for example `slsqp = SLSQP(optOptions=optOptions)` instead of `OPT('slsqp', optOptions=optOptions)`. As you said, we maintain a list of strings and map them to the optimizer [here](https://github.com/mdolab/pyoptsparse/blob/master/pyoptsparse/pyOpt_optimizer.py#L890), and I agree that this is not very elegant. Would it be sufficient to update the documentation/examples to recommend this approach instead of using strings?
