### 2020-10-05 7dc62b0
updated docstrings for utils
