### 2020-10-05 mdolab/pyoptsparse#169 @nwu63
Actually, I took a look at the code and I've reversed my stance. I think it's useful to be able to import a single object, i.e. `OPT` and change the optimizer on-the-fly. If I want to swap between two optimizers, it's much better to instantiate with `OPT(optimizer)` instead of wrapping everything in conditionals, essentially duplicating the code that is already [in pyOptSparse](https://github.com/mdolab/pyoptsparse/blob/master/pyoptsparse/pyOpt_optimizer.py#L890). That would unnecessarily clutter every runscript.
