### 2020-10-05 mdolab/pyoptsparse#169 @matbesancon
> Would it be sufficient to update the documentation/examples to recommend this approach instead of using strings?

in my opinion yes
