### 2020-10-05 b54bd35
Removed unnecessary comments for scipy (#171)
