### 2020-10-05 mdolab/pyoptsparse#169 @nwu63
Sorry Mathieu, not sure if you saw the comment above. I think there's value in being able to do this via strings since it allows you to quickly swap optimization algorithms. We also cast all the strings to lower case to try and reduce the possibility of typos.
