### 2020-10-05 mdolab/pyoptsparse#169 @dhill2522
Just an end user perspective, but could using an enum work well here? I think you could dodge the possibility of typos, while still allowing easily swappable optimizers.

I am thinking something along the lines of `opt = OPT(optimizers.IPOPT, optOptions=optOptions` where `optimizers` would be an enum defined in a module somewhere.
