### 2020-10-05 mdolab/pyoptsparse#169 @nwu63
Oh yeah that's a good idea. The capitalization of some optimizers is always tricky, for example `ParOpt` vs. `PAROPT`, but nevertheless I think this could work well. I'll play around with this and report back.
