### 2020-10-05 mdolab/pyoptsparse#169 @dhill2522
Sounds good. I can see that with the capitalization. Depending on the editor in use it may actually become easier as fancier editors will autocomplete enum values, but not strings.
