### 2020-10-06 mdolab/pyoptsparse#169 @matbesancon
> I think there's value in being able to do this via strings since it allows you to quickly swap optimization algorithms

yes I agree that it should remain an option. I raised this issue to wonder if it should be the principal way recommended in the documentation (included the quick start)
