### 2020-10-07 df5e7c2
additional edits based on feedback from reviewer
