### 2020-10-07 1736ae9
minor string fixes
