### 2020-10-07 3dc94a4
added enum as an option for OPT
