### 2020-10-08 c26cfa0
updated docs to explain two ways of creating optimizer
