### 2020-10-08 98790d8
updated examples to use nested dictionary for sens
