### 2020-10-08 3f8d38c
Update guide.rst
