### 2020-10-08 #2564 @nwu63
@matbesancon the issues you opened have been addressed, and I also pushed a minor update to the paper.

@poulson does JOSS require three reviewers? As far as I can tell @jgoldfar has not started his review yet.
