### 2020-10-08 4bede39
updated var plot script
