+++ examples/hs015VarPlot.py
@@ -1,56 +1,58 @@
+"""
+This script shows a possible post-processing approach once the history file is generated.
+It visualizes the design space, and plots the objective and constraint contours.
+It also reads in the history files and plots the path of the optimization.
+
+In order to visualize the optimization history, the history files must be available.
+They can be generated by running the hs015.py script, e.g.
+
+>>> python hs015.py --storeHistory --opt slsqp
+"""
+
 import numpy as np
 import matplotlib.pyplot as plt
-from sqlitedict import SqliteDict
+from pyoptsparse import History
 
-db = SqliteDict()
+db = {}
 opts = ["ipopt", "slsqp", "snopt", "conmin", "nlpqlp", "psqp"]
 for opt in opts:
-    fileName = "%s_hs015_Hist.hst" % opt
+    fileName = "{}_hs015_Hist.hst".format(opt)
     try:
-        db[opt] = SqliteDict(fileName)
-    except:  # noqa: E722
+        db[opt] = History(fileName)
+    except FileNotFoundError:
         pass
 
 obj = {}
-x1 = {}
-x2 = {}
+xuser = {}
 
 for opt in db.keys():
-    n = int(db[opt]["last"])
-
-    obj[opt] = []
-    x1[opt] = []
-    x2[opt] = []
-    for i in range(n):
-        try:
-            obj[opt].append(db[opt]["%d" % i]["funcs"]["obj"])
-            x1[opt].append(db[opt]["%d" % i]["xuser"]["xvars"][0])
-            x2[opt].append(db[opt]["%d" % i]["xuser"]["xvars"][1])
-        except:  # noqa: E722
-            pass
+    val = db[opt].getValues()
+    obj[opt] = val["obj"]
+    xuser[opt] = val["xvars"]
 
 # Generate the Rosenbrock contours
-delta = 0.25
+delta = 0.1
 x = np.arange(-2.5, 1.5, delta)
 y = np.arange(-6.5, 3.0, delta)
 X, Y = np.meshgrid(x, y)
 
+# objective
 Z = 100 * (Y - X ** 2) ** 2 + (1 - X) ** 2
 # and the constraint contours
 A = X * Y
 B = X + Y ** 2
 
 # plot the contours and constraints
-plt.figure(figsize=(11, 8.5))
-levels = [6000, 3000, 2000, 1000, 500, 250]
+plt.figure()
+levels = [250, 500, 1000, 2000, 3000, 6000]
 CS = plt.contour(X, Y, Z, levels, colors="k")
 levels = np.arange(1.0, 1.01)
 CS1 = plt.contour(X, Y, A, levels, colors="g")
 levels = np.arange(0.0, 0.01)
 CS2 = plt.contour(X, Y, B, levels, colors="b")
-plt.clabel(CS, inline=1, fontsize=10)
-plt.clabel(CS1, inline=1, fontsize=10)
-plt.clabel(CS2, inline=1, fontsize=10)
+plt.clabel(CS, inline=1)
+plt.clabel(CS1, inline=1)
+plt.clabel(CS2, inline=1)
 
 # set the one sided variable
 xupper = [0.5, 0.5]
@@ -60,10 +62,9 @@ yupper = [-7, 3.0]
 styleList = ["ko-", "ro-", "bo-", "go-", "mo-", "co-", "ks--"]
 counter = 0
 for opt in db.keys():
-    plt.plot(x1[opt], x2[opt], styleList[counter], label="%s" % (opt))
+    plt.plot(xuser[opt][:, 0], xuser[opt][:, 1], styleList[counter], label=opt)
     counter += 1
 
-# end
 plt.plot(xupper, yupper, "k")
 plt.legend(loc=3)
 plt.xlabel("x1")
