### 2020-10-08 db5ee4b
remove unhelpful examples
