### 2020-10-08 7c5c3bc
updated rosenbrock example
