+++ examples/rosenbrock.py
@@ -1,26 +1,13 @@
 #!/usr/bin/env python
 from pyoptsparse import Optimization, OPT
 import argparse
-import numpy as np
-
-
-# import cProfile
 
 parser = argparse.ArgumentParser()
-parser.add_argument("--sens", help="sensitivity mode", type=str, default="FD")
-parser.add_argument("--constrained", help="constrained or not", type=int, default=0)
-parser.add_argument("--testHist", help="test history", type=str, default="no")
-parser.add_argument("--groups", help="use groups", type=int, default=0)
-parser.add_argument("--sensMode", help="gradient mode", type=str, default="")
+parser.add_argument("--sens", help="sensitivity mode", type=str, default="FD", choices=["FD", "CS", "CD", "user"])
+parser.add_argument("--constrained", help="add a constraint", action="store_true")
+parser.add_argument("--testHotstart", help="test hotstart", action="store_true")
 parser.add_argument("--opt", help="optimizer", type=str, default="SLSQP")
-parser.add_argument("--pythonProf", help="profile in Python", type=int, default=0)
 args = parser.parse_args()
-sens = args.sens
-constrained = args.constrained
-testHist = args.testHist
-groups = args.groups
-sensMode = args.sensMode
-pythonProf = args.pythonProf
 optOptions = {}
 
 
@@ -45,87 +32,50 @@ def sensfunc(xdict, funcsDict):
     return funcsSens, fail
 
 
-# Matrix-free algorithm return functions
-def objgrad(xdict):
-    x = xdict["xvars"]
-
-    objsens = {}
-    objsens["obj"] = {
-        "xvars": np.array([2 * 100 * (x[1] - x[0] ** 2) * (-2 * x[0]) - 2 * (1 - x[0]), 2 * 100 * (x[1] - x[0] ** 2)])
-    }
-    fail = False
-
-    return objsens, fail
-
-
-def jprod(xdict, pdict, sparse_only):
-    x = xdict["xvars"]
-    p = pdict["xvars"]
-    qdict = {}
-
-    if constrained:
-        gcon = np.array([-3 * (x[0] - 1) ** 2, -1])
-        qdict["con"] = np.array([np.dot(gcon, p)])
-    fail = False
-
-    return qdict, fail
-
-
-def jtprod(xdict, qdict, sparse_only):
-    x = xdict["xvars"]
-
-    pdict = {}
-    if constrained:
-        q = qdict["con"]
-        gcon = np.array([-3 * (x[0] - 1) ** 2, -1])
-        pdict["xvars"] = q * gcon
-    else:
-        pdict["xvars"] = np.zeros(len(x))
-    fail = False
-
-    return pdict, fail
-
-
-if sens == "none":
-    sens = None
-if sens == "user":
+if args.sens == "user":
     sens = sensfunc
-if sens == "matrix-free":
-    sens = [objgrad, jprod, jtprod]
+else:
+    sens = args.sens
 
 # Instantiate Optimization Problem
 optProb = Optimization("Rosenbrock function", objfunc)
 optProb.addVarGroup("xvars", 2, "c", value=[3, -3], lower=-5.12, upper=5.12, scale=[1.0, 1.0])
-optProb.finalizeDesignVariables()
-if constrained:
+if args.constrained:
     optProb.addCon("con", upper=0, scale=1.0)
 optProb.addObj("obj")
 
 # Create optimizer
 opt = OPT(args.opt, options=optOptions)
-if testHist == "no":
-    # Just run a normal run
-    sol = opt(optProb, sens=sens, sensMode=sensMode)
-    # print(sol.fStar)
-    print(sol)
-else:
+
+if args.testHotstart:
+    histName = "opt_hist.hst"
[…]
