+++ examples/hs015VarPlot.py
@@ -4,9 +4,10 @@ It visualizes the design space, and plots the objective and constraint contours.
 It also reads in the history files and plots the path of the optimization.
 
 In order to visualize the optimization history, the history files must be available.
-They can be generated by running the hs015.py script, e.g.
+They can be generated by running the hs015.py script for each optimizer, e.g.
 
 >>> python hs015.py --storeHistory --opt slsqp
+>>> python hs015.py --storeHistory --opt conmin
 """
 
 import numpy as np
