### 2020-10-08 3e1f8f5
made rosenbrock example only run SLSQP
