+++ paper/paper.bib
@@ -1,169 +1,170 @@
 @article{Perez2012,
-    author = {Ruben E. Perez and Peter W. Jansen and Joaquim R. R. A. Martins},
-    doi = {10.1007/s00158-011-0666-3},
+    author  = {Ruben E. Perez and Peter W. Jansen and Joaquim R. R. A. Martins},
+    doi     = {10.1007/s00158-011-0666-3},
     journal = {Structural and Multidisciplinary Optimization},
-    month = {January},
-    number = {1},
-    pages = {101--118},
-    title = {{pyOpt}: A {Python}-Based Object-Oriented Framework for Nonlinear Constrained Optimization},
-    volume = {45},
-    year = {2012}
+    month   = {January},
+    number  = {1},
+    pages   = {101--118},
+    title   = {{pyOpt}: A {Python}-Based Object-Oriented Framework for Nonlinear Constrained Optimization},
+    volume  = {45},
+    year    = {2012}
 }
 
 @article{Gill2005,
-    author = {Philip E. Gill and Walter Murray and Michael A. Saunders},
-    doi = {10.1137/S0036144504446096},
-    journal = {SIAM Review},
-    number = {1},
-    pages = {99-131},
+    author    = {Philip E. Gill and Walter Murray and Michael A. Saunders},
+    doi       = {10.1137/S0036144504446096},
+    journal   = {SIAM Review},
+    number    = {1},
+    pages     = {99-131},
     publisher = {SIAM},
-    title = {{SNOPT}: An {SQP} Algorithm for Large-Scale Constrained Optimization},
-    volume = {47},
-    year = {2005}
+    title     = {{SNOPT}: An {SQP} Algorithm for Large-Scale Constrained Optimization},
+    volume    = {47},
+    year      = {2005}
 }
 
 @article{Schittkowski2006,
-    title={NLPQLP: A fortran implementation of a sequential quadratic programming algorithm with distributed and non-monotone line search--user’s guide},
-    author={Schittkowski, Klaus},
-    journal={Report, Department of Computer Science, University of Bayreuth},
-    year={2006}
+    title   = {NLPQLP: A fortran implementation of a sequential quadratic programming algorithm with distributed and non-monotone line search--user’s guide},
+    author  = {Schittkowski, Klaus},
+    journal = {Report, Department of Computer Science, University of Bayreuth},
+    year    = {2006}
 }
 
 @techreport{Kraft1988,
-    author = {Kraft, D},
+    author      = {Kraft, D},
     institution = {Tech. Rep. DFVLR-FB 88-28, DLR German Aerospace Center},
-    title = {{A software package for sequential quadratic programming}},
-    url = {http://www.opengrey.eu/item/display/10068/147127},
-    year = {1988}
+    title       = {{A software package for sequential quadratic programming}},
+    url         = {http://www.opengrey.eu/item/display/10068/147127},
+    year        = {1988}
 }
 
 @article{Curtis2010,
-    author = {Frank E. Curtis and Olaf Schenk and Adreas W\"achter},
+    author  = {Frank E. Curtis and Olaf Schenk and Adreas W\"achter},
     journal = {SIAM Journal on Scientific Computing},
-    number = {6},
-    pages = {3447--3475},
-    title = {An Interior Point Algorithm for Large-Scale Nonlinear Optimization with Inexact Step Computations},
-    volume = {32},
-    year = {2010}
+    number  = {6},
+    pages   = {3447--3475},
[…]
