### 2020-10-10 3333a5b
more bib updates
