+++ paper/paper.bib
@@ -10,86 +10,6 @@
     year    = {2012}
 }
 
-@article{Gill2005,
-    author    = {Philip E. Gill and Walter Murray and Michael A. Saunders},
-    doi       = {10.1137/S0036144504446096},
-    journal   = {SIAM Review},
-    number    = {1},
-    pages     = {99-131},
-    publisher = {SIAM},
-    title     = {{SNOPT}: An {SQP} Algorithm for Large-Scale Constrained Optimization},
-    volume    = {47},
-    year      = {2005}
-}
-
-@article{Schittkowski2006,
-    title   = {NLPQLP: A fortran implementation of a sequential quadratic programming algorithm with distributed and non-monotone line search--user’s guide},
-    author  = {Schittkowski, Klaus},
-    journal = {Report, Department of Computer Science, University of Bayreuth},
-    year    = {2006}
-}
-
-@techreport{Kraft1988,
-    author      = {Kraft, D},
-    institution = {Tech. Rep. DFVLR-FB 88-28, DLR German Aerospace Center},
-    title       = {{A software package for sequential quadratic programming}},
-    url         = {http://www.opengrey.eu/item/display/10068/147127},
-    year        = {1988}
-}
-
-@article{Curtis2010,
-    author  = {Frank E. Curtis and Olaf Schenk and Adreas W\"achter},
-    journal = {SIAM Journal on Scientific Computing},
-    number  = {6},
-    pages   = {3447--3475},
-    title   = {An Interior Point Algorithm for Large-Scale Nonlinear Optimization with Inexact Step Computations},
-    volume  = {32},
-    year    = {2010}
-}
-
-@article{Vanderplaats1973,
-    title  = {CONMIN: a Fortran program for constrained function minimization: user's manual},
-    author = {Vanderplaats, Garret N},
-    year   = {1973}
-}
-
-@article{Jansen2011,
-    doi     = {10.1016/j.compstruc.2011.03.011},
-    isbn    = {0045-7949},
-    issn    = {00457949},
-    journal = {Computers {\&} Structures},
-    number  = {13-14},
-    pages   = {1352--1366},
-    title   = {{Constrained structural design optimization via a parallel augmented Lagrangian particle swarm optimization approach}},
-    url     = {http://www.sciencedirect.com/science/article/pii/S0045794911000824},
-    volume  = {89},
-    year    = {2011}
-}
-
-@article{Deb2002,
-    author  = {Deb, K. and Pratap, A. and Agarwal, S. and Meyarivan, T.},
-    doi     = {10.1109/4235.996017},
-    issn    = {1089-778X},
-    journal = {IEEE Transactions on Evolutionary Computation},
-    month   = {apr},
-    number  = {2},
-    pages   = {182--197},
-    title   = {A fast and elitist multiobjective genetic algorithm: {NSGA-II}},
-    volume  = {6},
-    year    = {2002}
-}
-
-@article{Chin2019,
-    author  = {Ting Wei Chin and Mark L Leader and Graeme J Kennedy},
-    date    = {2019-09-01},
-    doi     = {https://doi.org/10.1016/j.advengsoft.2019.05.004},
-    journal = {Advances in Engineering Software},
-    title   = {A Scalable Framework for Large-Scale 3D Multimaterial Topology Optimization with Octree-based Mesh Adaptation},
-    url     = {http://gkennedy.gatech.edu/wp-content/uploads/2019/07/Chin_etal_2019_Multimaterial_octree_adapt_compressed.pdf},
-    volume  = {135},
-    year    = {2019}
-}
-
 @article{Brooks2018,
[…]
