### 2020-10-14 #2564 @matbesancon
@poulson I finished my review. This is a great paper covering a very interesting software package. The article highlights the interest optimization practitionners can have using it compared to other modelling libraries. I recommend its acceptation for JOSS
