+++ doc/conf.py
@@ -1,8 +1,4 @@
-# Configuration file for the Sphinx documentation builder.
-#
-# This file only contains a selection of the most common options. For a full
-# list see the documentation:
-# https://www.sphinx-doc.org/en/master/usage/configuration.html
+from sphinx_mdolab_theme.config import *
 
 # -- Path setup --------------------------------------------------------------
 
@@ -19,50 +15,11 @@ sys.path.insert(0, os.path.abspath("./"))  # to import custom Sphinx extension
 # -- Project information -----------------------------------------------------
 
 project = "pyOptSparse"
-copyright = "2020, MDO Lab"
-
-# The full version, including alpha/beta/rc tags
-# release = 'version'
-
 
 # -- General configuration ---------------------------------------------------
-
 
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = [
-    "sphinx.ext.autodoc",
-    "sphinx.ext.intersphinx",
-    "sphinx.ext.viewcode",
-    "numpydoc",
-    "sphinx.ext.autosummary",
-    "sphinx.ext.mathjax",
-    "ext.optimizertable",
-]
+extensions.extend(["numpydoc", "ext.optimizertable"])
 numpydoc_show_class_members = False
-
-# tell autodoc to document the __init__ methods
-autoclass_content = "both"
-
-# Add any paths that contain templates here, relative to this directory.
-templates_path = ["_templates"]
-
-# List of patterns, relative to source directory, that match files and
-# directories to ignore when looking for source files.
-# This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ["_build"]
-
-# The master toctree document.
-master_doc = "index"
-
-# -- Options for HTML output -------------------------------------------------
-
-# The theme to use for HTML and HTML Help pages.  See the documentation for
-# a list of builtin themes.
-html_theme = "sphinx_rtd_theme"
-
-# Add any paths that contain custom static files (such as style sheets) here,
-# relative to this directory. They are copied after the builtin static files,
-# so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ["_static"]
