### 2020-10-19 7b762b2
Update .travis.yml
