+++ .travis.yml
@@ -28,7 +28,7 @@ script:
   - |
     if [ $TEST_TYPE == "docker" ]; then
       # run tests
-      docker exec -it app /bin/bash -c ". $BASHRC && cd $DOCKER_WORKING_DIR/test && testflo --pre_announce -v . --coverage --coverpkg $REPO_NAME";
+      docker exec -it app /bin/bash -c ". $BASHRC && cd $DOCKER_WORKING_DIR/test && testflo --pre_announce -v . --coverage --coverpkg $REPO_NAME"
     fi
 
 after_success:
