+++ .travis.yml
@@ -8,7 +8,7 @@ env:
     - DOCKER_WORKING_DIR=/home/mdolabuser/packages/$REPO_NAME # pyOptSparse is under packages, so this should overwrite what's set in travis-env.yml
     - SNOPT_DIR=$DOCKER_WORKING_DIR/$REPO_NAME/pySNOPT/source
     - NLPQLP_DIR=$DOCKER_WORKING_DIR/$REPO_NAME/pyNLPQLP/source
-    - AUTO_DOCKER_REPO=true
+    - AUTO_DOCKER_REPO=true # this will automatically set DOCKER_REPO depending on where the build originates
 
 install:
   # We back up the proprietary source codes first
