+++ .travis.yml
@@ -36,3 +36,4 @@ after_success:
   # These environment variables must be exposed to coveralls for it to work (since we are using Docker)
   - docker exec -it app /bin/bash -c ". $BASHRC && pip install coveralls";
   - docker exec -e COVERALLS_REPO_TOKEN="$COVERALLS_REPO_TOKEN" -e TRAVIS_JOB_ID="$TRAVIS_JOB_ID" -e TRAVIS_BRANCH="$TRAVIS_BRANCH" -it app /bin/bash -c ". \$HOME/.bashrc_mdolab && cd $DOCKER_WORKING_DIR && coveralls --rcfile=.coveragerc";
+
