### 2020-10-19 #2564 @poulson
Hi @nwu63 -- with respect to your three reviewer question, given that there are two very positive reviews I think that is sufficient.

I will procede by giving it a proofread and then walk through the checklist for the final approval.

Thank you all for your patience, as I fell into a bit of a spiral of knowing I was behind on checking in and avoiding circling back.
