### 2020-10-19 #2564 @poulson
@nwu63 -- I just reread the paper, and I have to say, this is one of my favorite JOSS submissions.

Would you mind fixing the trivial DOI issue mentioned above and then providing a DOI for your source code archive (e.g., from Zenodo)?
