### 2020-10-20 #2564 @nwu63
@poulson: Thanks! This has been a really good experience and I appreciate all the work you guys do.

I am planning to make a new code release so that the updated code is archived by zenodo. I just wanted to clarify if the code release should contain the paper? i.e. should I merge the content of the `paper` branch into `master` before that release. I saw from some other JOSS reviews that you prefer having the zenodo metadata match the paper, so I will add the json metadata file to remedy that as well.
