### 2020-10-20 #2564 @nwu63
@poulson I have just made a new release which also merged the contents of the paper. The zenodo archive is also there now with the correct author list, the DOI is `10.5281/zenodo.4110792`.
