### 2020-10-20 fa97c6b
Update Travis (#168)
* Update .travis.yml

* Delete main.yml

* Update .flake8

* Update .travis.yml

* Update .travis.yml

* Update .travis.yml

* Update .travis.yml

* Update .travis.yml

* Update .travis.yml

* Update .travis.yml

* Update .travis.yml

* Update .travis.yml

* Update .travis.yml

* restored .flake8

* Update .travis.yml

* Update .travis.yml

* Update .travis.yml

* Update .travis.yml

* Update .travis.yml

* Update .travis.yml

* Update .travis.yml

* Update .travis.yml

* Update .travis.yml

* Update .flake8

* Update .travis.yml

* Update .travis.yml

* Update .travis.yml
[…]
