+++ .travis.yml
@@ -1,50 +1,38 @@
 import:
   - mdolab/.github:travis/jobs-env.yml
+  - mdolab/.github:travis/before-install.yml # we skip importing the install step because this repo requires extra steps
+  - mdolab/.github:travis/script.yml
 
 env:
   global:
     - DOCKER_WORKING_DIR=/home/mdolabuser/packages/$REPO_NAME # pyOptSparse is under packages, so this should overwrite what's set in travis-env.yml
     - SNOPT_DIR=$DOCKER_WORKING_DIR/$REPO_NAME/pySNOPT/source
     - NLPQLP_DIR=$DOCKER_WORKING_DIR/$REPO_NAME/pyNLPQLP/source
-
-before_install:
-  - if [ $PR_FROM != "mdolab" ]; then
-      export DOCKER_REPO=public;
-    else
-      export DOCKER_REPO=private;
-      echo "$DOCKER_PASSWORD" | docker login -u "$DOCKER_USERNAME" --password-stdin;
-    fi
-  - docker pull mdolab/$DOCKER_REPO:$DOCKER_TAG
-  # run Docker, key is we mount the current Travis directory into Docker to access content of repo
-  - docker run -t -d --rm
-        --name app
-        --mount "type=bind,src=$(pwd),target=$DOCKER_MOUNT_DIR"
-        mdolab/$DOCKER_REPO:$DOCKER_TAG
-        /bin/bash
+    - AUTO_DOCKER_REPO=true # this will automatically set DOCKER_REPO depending on where the build originates
 
 install:
   # We back up the proprietary source codes first
-  - if [ $PR_FROM == "mdolab" ]; then
+  - if [ $DOCKER_REPO == "private" ]; then
       docker exec -it app /bin/bash -c "cp -r $SNOPT_DIR \$HOME/SNOPT && cp -r $NLPQLP_DIR \$HOME/NLPQLP";
     fi
   # We thrown away the existing repo in Docker, and copy the new one in-place
   - docker exec -it app /bin/bash -c "rm -rf $DOCKER_WORKING_DIR && cp -r $DOCKER_MOUNT_DIR $DOCKER_WORKING_DIR"
-  # We also remove the python installation
-  - docker exec -it app /bin/bash -c "rm -rf $DOCKER_WORKING_DIR/build"
   # Copy back the proprietary codes
-  - if [ $PR_FROM == "mdolab" ]; then
+  - if [ $DOCKER_REPO == "private" ]; then
       docker exec -it app /bin/bash -c "cp -r \$HOME/NLPQLP/* $NLPQLP_DIR/ && cp -r \$HOME/SNOPT/* $SNOPT_DIR/";
     fi
   # Build and install
-  - docker exec -it app /bin/bash -c ". $BASHRC && cd $DOCKER_WORKING_DIR && pip install -e ."
+  - docker exec -it app /bin/bash -c ". $BASHRC && cd $DOCKER_WORKING_DIR && pip install ."
+
 script:
-  # We need to source the mdolab bashrc before running anything
-  - docker exec -it app /bin/bash -c ". $BASHRC && cd $DOCKER_WORKING_DIR && testflo --pre_announce -v . --coverage --coverpkg $REPO_NAME"
+  - |
+    if [ $TEST_TYPE == "docker" ]; then
+      # run tests
+      docker exec -it app /bin/bash -c ". $BASHRC && cd $DOCKER_WORKING_DIR/test && testflo --pre_announce -v . --coverage --coverpkg $REPO_NAME"
+    fi
 
 after_success:
   # Coveralls
[…]
