+++ .github/workflows/main.yml
@@ -1,37 +1,0 @@
-name: codestyle-checks
-on:
-  push:
-    branches: [ master ]
-  pull_request:
-    branches: [ master ]
-
-jobs:
-  black-format:
-    runs-on: ubuntu-latest
-    strategy:
-      fail-fast: false
-      matrix:
-        python-version: [ '37', '38' ]
-    steps:
-    - uses: actions/checkout@v2
-    - name: Black Code Formatter
-      uses: lgeiger/black-action@v1.0.1
-      with:
-          args: ". --check -l 120 --target-version py${{ matrix.python-version }} --exclude '/setup.py|doc/'"
-  flake8:
-    runs-on: ubuntu-latest
-    strategy:
-      fail-fast: false
-      matrix:
-        python-version: [ '3.7', '3.8' ]
-    name: flake8 with Python ${{ matrix.python-version }}
-    steps:
-    - uses: actions/checkout@v2
-    - name: Set up Python
-      uses: actions/setup-python@v2
-      with:
-        python-version: ${{ matrix.python-version }}
-    - name: Install dependencies
-      run: pip install flake8
-    - name: Lint with flake8
-      run: flake8 . --count --show-source --statistics
