### 2020-10-24 #2564 @nwu63
Hi @kyleniemeyer and everyone else involved in this, thanks so much! Just one quick question: there is an incorrect capitalization in the author list [here](https://joss.theoj.org/papers/10.21105/joss.02564), `Joaquim R. r. a. Martins` should be `Joaquim R. R. A. Martins`, is there a way to fix that?
