### 2020-10-25 #2564 @arfon
> @openjournals/dev can you help with this? the issue is in the author list on the landing page—the paper itself is fine

Fixed.
